% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_email_content.R
\name{build_email_content}
\alias{build_email_content}
\title{Merge Message Contents into Single Document}
\usage{
build_email_content()
}
\value{

}
\description{
Merge Message Contents into Single Document
}
