% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_df.R
\name{create_summary_df}
\alias{create_summary_df}
\title{Create a summary dataframe for use in a table}
\usage{
create_summary_df(data, years_in_sparklines = 3, pretty_round = TRUE)
}
\arguments{
\item{data}{Dataframe of input data}

\item{years_in_sparklines}{Number of years worth of data to include in
sparkline}

\item{pretty_round}{Logical; \code{TRUE} by default. If \code{TRUE}, figures will be
rounded using \code{pretty_round()}. This means (for eg.) figures below 50
will be rounded to 0. If \code{FALSE}, figures will not be rounded with
\code{pretty_round()}.}
}
\description{
Create a summary dataframe for use in a table
}
\examples{
\dontrun{
dash_data <- load_dash_data()
create_summary_df(data = filter_dash_data(c(
  "A84423349V",
  "A84423357V",
  "A84423356T",
  "A84423244X",
  "A84423468K",
  "pt_emp_vic"
)))
}
}
