% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_email_content.R
\name{edit_email_lists}
\alias{edit_email_lists}
\title{Edit Distribution Lists}
\usage{
edit_email_lists()
}
\value{

}
\description{
Edit Distribution Lists
}
\examples{
\dontrun{
edit_email_lists()
}
}
