% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_email.R
\name{editEMAIL}
\alias{editEMAIL}
\title{Edit SMS before Sending}
\usage{
editEMAIL()
}
\arguments{
\item{labour}{data.frame used for testing, not required by addin}
}
\description{
Edit SMS before Sending
}
