% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_sms.R
\name{editSMS}
\alias{editSMS}
\title{Edit SMS before Sending}
\usage{
editSMS(labour = NULL)
}
\arguments{
\item{labour}{data.frame used for testing, not required by addin}
}
\value{

}
\description{
Edit SMS before Sending
}
