% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_plot.R
\name{example_plot}
\alias{example_plot}
\title{An example function that creates a plot}
\usage{
example_plot(data)
}
\description{
An example function that creates a plot
}
\examples{
\dontrun{
dash_data <- load_dash_data()
df <- filter_dash_data(c("A84423242V", "A84423466F"), df = dash_data)
example_plot(df)
}

}
