% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dash_data.R
\name{filter_dash_data}
\alias{filter_dash_data}
\title{Convenience function to extract specified ABS series IDs from \code{dash_data}}
\usage{
filter_dash_data(series_ids, df = dash_data)
}
\arguments{
\item{series_ids}{Vector of series ID(s)}

\item{df}{Data frame, expected to be the df returned by \code{load_dash_data()}}
}
\value{
An unnested tbl_df containing only the specified series ID(s)
}
\description{
Convenience function to extract specified ABS series IDs from \code{dash_data}
}
\examples{
\dontrun{
dash_data <- load_dash_data()
filter_dash_data("A84423354L")
}
}
