% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_update-data-and-knit.R
\name{generate_report}
\alias{generate_report}
\title{Knit Report to Word}
\usage{
generate_report()
}
\value{

}
\description{
This Rstudio addin will run the code to render the Monthly report.
}
