% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dash_data.R
\name{get_dash_data}
\alias{get_dash_data}
\title{Get dash data table}
\usage{
get_dash_data(
  refresh = c("auto", "always", "never"),
  branch = "main",
  verbose = F
)
}
\arguments{
\item{refresh}{Refresh behaviour. Possible values:
\itemize{
\item \code{auto}: will re-download data if it is out of date
\item \code{always}: will always re-download data
\item \code{never}: will never re-download data
}}

\item{branch}{The GitHub branch to download data from}

\item{verbose}{Whether to print verbose messages}
}
\value{
A table with dashboard data.

If \code{refresh} was \code{auto} or \code{always}, \verb{attr(*, "date_updated")} will contain
a timestamp of when the data was last updated.
get_dash_data(refresh = "never") # Just return internally cached data
}
\description{
By default, this will re-download data from the \code{djprdashdata} repo if
the current dashboard's data is out of date.
}
