% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dash_data.R
\name{get_summ}
\alias{get_summ}
\title{Get a key fact about a time series}
\usage{
get_summ(series_ids, item, df = ts_summ)
}
\arguments{
\item{series_ids}{Character vector of ABS time series ID(s)}

\item{item}{Item to extract from summary; see \code{?ts_summarise()}
for the full list. Unquoted.}

\item{df}{Data frame containing time series summarised with
\code{djprshiny::ts_summarise()}.}
}
\description{
Get a key fact about a time series
}
\examples{
\dontrun{
dash_data <- load_and_hide()

ts_summ <- dash_data \%>\%
  tidyr::unnest(cols = data) \%>\%
  djprshiny::ts_summarise()

get_summ("A84601638A", latest_value)
}

}
