% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_briefing.R
\name{knit_briefing}
\alias{knit_briefing}
\title{Create a Word document containing briefing tables
This function knits a briefing document that uses tables generated in this
package to create a Word document. It knits the RMarkdown file
\code{jobs_briefing.Rmd} that lives in the \code{inst} folder of \code{djprlabourdash}.}
\usage{
knit_briefing(path = here::here("inst"), quietly = FALSE)
}
\arguments{
\item{path}{Path to directory where
the knitted Word document should be created.}

\item{quietly}{Passed to \code{rmarkdown::render()}'s \code{quiet} argument. \code{FALSE} by
default.}
}
\description{
Create a Word document containing briefing tables
This function knits a briefing document that uses tables generated in this
package to create a Word document. It knits the RMarkdown file
\code{jobs_briefing.Rmd} that lives in the \code{inst} folder of \code{djprlabourdash}.
}
\examples{
\dontrun{
knit_briefing(quietly = TRUE)
}
}
