% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dash_data.R
\name{load_dash_data}
\alias{load_dash_data}
\title{Load data for the DJPR Labour Dashboard}
\usage{
load_dash_data(branch = "main")
}
\arguments{
\item{branch}{Github branch of \code{djprdashdata} repo from which to load dash data}
}
\description{
Load data for the DJPR Labour Dashboard
}
