% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_briefing_table.R
\name{make_briefing_table}
\alias{make_briefing_table}
\title{Take a data frame summarising data and create a flextable for
a briefing table}
\usage{
make_briefing_table(
  df,
  highlight_rows = NULL,
  header_row = c("", "Current figures", "Change in past month", "Change in past year",
    "Change during govt")
)
}
\arguments{
\item{df}{Dataframe created using \code{create_summary_df()}}

\item{highlight_rows}{numeric vector of rows in the table body (ie.
excluding column names / header row) to highlight. Non-highlighted rows
will be indented. When \code{NULL}, no row is highlighted.}

\item{header_row}{Character vector of names of additional header row to be
added above existing column names}
}
\value{
A \code{flextable::flextable} object
}
\description{
Take a data frame summarising data and create a flextable for
a briefing table
}
