% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_reactable.R
\name{make_reactable}
\alias{make_reactable}
\title{Make a reactable with standard formatting}
\usage{
make_reactable(
  summary_df,
  raw_data,
  years_in_sparklines = 2,
  highlight_rows = NULL
)
}
\arguments{
\item{summary_df}{a data frame, ideally created by \code{create_summary_df()}}

\item{raw_data}{data frame from which \code{summary_df} was created}

\item{years_in_sparklines}{number of years prior to latest obs to include
in sparkline}

\item{highlight_rows}{numeric vector of rows in the table body (ie.
excluding column names / header row) to highlight. Non-highlighted rows
will be indented. When \code{NULL}, no row is highlighted.}
}
\description{
Each row is an indicator; columns are levels / change over particular periods
Conditional formatting is applied, sparklines are included.
}
