% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sparklines.R
\name{make_sparklines}
\alias{make_sparklines}
\title{Function that takes a dataframe and returns a list of line charts}
\usage{
make_sparklines(df, group_var, date_var = date, value_var = value)
}
\arguments{
\item{df}{Dataframe containing data to summarise in sparkline form}

\item{group_var}{Unquoted name of grouping variable in \code{df}. One line chart
will be created for each unique value in \code{group_var}.}

\item{date_var}{Unquoted name of date variable in \code{df} to plot on the
x-axis in the line charts. Default is \code{date}.}

\item{value_var}{Unquoted name of variable in \code{df} to map to the y aesthetic
of the line charts. Default is \code{value}.}
}
\value{
A list of line charts (\code{{ggplot2}} objects).
}
\description{
One line chart is returned for each level in the grouping variable. These
line charts are intended for use as sparklines in tables.
}
\details{
All line charts in the list will have the same maximum and minimum
dates on the x-axis, defined by the range in the \code{date_var} in \code{df}.
}
\examples{
make_sparklines(df = ggplot2::economics_long, group_var = variable)
}
