% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\title{Create a table for the dashboard or a briefing document}
\usage{
make_table(
  data,
  destination = Sys.getenv("R_DJPRLABOURDASH_TABLEDEST", unset = "dashboard"),
  years_in_sparklines = 3,
  row_order = NULL,
  highlight_rows = NULL,
  notes = NULL,
  title = "",
  rename_indicators = TRUE
)
}
\arguments{
\item{data}{A data frame containing data to summarise}

\item{destination}{"dashboard" or "briefing"}

\item{years_in_sparklines}{Period of time to include in the sparkline line
charts.}

\item{row_order}{Vector of series IDs, in the order in which you wish the
corresponding rows to be included in the output table}

\item{highlight_rows}{Vector of series IDs, corresponding to rows
in the table to highlight.
Highlighted rows are bolded and have a top border; non-highlighted rows
are indented. If \code{NULL} then all rows are non-bold, non-indented.}

\item{notes}{Optional notes to add to caption. Source will be inferred
automatically based on the data using \code{caption_auto()}.}

\item{title}{Character vector to use as the table title. Will only be used
when \code{destination} is "briefing".}

\item{rename_indicators}{logical; default is \code{TRUE}. If \code{TRUE}, the
\code{rename_indicators()} function will be used to rename certain indicators.}
}
\description{
Create a table for the dashboard or a briefing document
}
\examples{
# dash_data <- load_dash_data()
\dontrun{
make_table(
  data = filter_dash_data(series_ids = c(
    "A84423354L",
    "A84423242V",
    "A84423466F"
  )),
  row_order = c(
    "A84423354L",
    "A84423242V",
    "A84423466F"
  ),
  highlight_rows = c("A84426256L")
)
}
}
