% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_round.R
\name{pretty_round}
\alias{pretty_round}
\title{Function to return numbers rounded for presentation purposes,
with commas where appropriate.}
\usage{
pretty_round(x)
}
\arguments{
\item{x}{Numeric vector to round}
}
\value{
Character vector of rounded numbers, with commas separating
thousands and 'm' to denote millions.

Numbers below 1m are rounded to the nearest 100. Numbers >= 1m are
rounded to the nearest 1,000.
}
\description{
Function to return numbers rounded for presentation purposes,
with commas where appropriate.
}
\examples{
pretty_round(c(10, 110, 10050, 20500, 2000100, 3445123))
}
