% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_indicators.R
\name{rename_indicators}
\alias{rename_indicators}
\title{Convenience function to give more descriptive names to specific indicators}
\usage{
rename_indicators(data)
}
\arguments{
\item{data}{A dataframe containing an 'indicator' column and a 'series_id'
column.}
}
\value{
A dataframe, with the indicator column modified if any specific
series IDs are encountered.
}
\description{
Intended for use with \code{make_table()}.
}
\details{
This function looks for specific values in the \code{series_id} column; if encountered,
the \code{indicator} column in the row is modified.
}
\keyword{internal}
