% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_years.R
\name{subtract_years}
\alias{subtract_years}
\title{Subtract a number of years from a date}
\usage{
subtract_years(date, n_years)
}
\arguments{
\item{date}{Date from which to subtract \code{n_years}}

\item{n_years}{Number of years to subtract. Will be coerced to integer -
cannot subtract fractions of a year.}
}
\description{
Subtract a number of years from a date
}
\examples{
subtract_years(Sys.Date(), 5)
subtract_years(as.Date("2020-01-01"), 1)
\dontrun{
# This returns an error:
subtract_years(Sys.Date(), 1.5)
}
}
