% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{table_overview}
\alias{table_overview}
\title{Create a Reactable to summarise key ABS LFS indicators}
\usage{
table_overview(
  data = filter_dash_data(series_ids = c("A84423349V", "A84423355R", "A84423354L",
    "A84423350C", "A85223451R", "A84426256L", "A85223450L", "A84423357V", "pt_emp_vic")),
  years_in_sparklines = 2,
  row_var = "indicator"
)
}
\arguments{
\item{data}{A dataframe containing tidied ABS Labour Force Survey}

\item{years_in_sparklines}{Numeric; indicating the number of years of data
to include in the sparkline charts in the returned reactable.}

\item{row_var}{String - variable name in data to use for row names,
as in "indicator"}
}
\value{
A reactable (htmlwidget) including sparklines
}
\description{
Create a Reactable to summarise key ABS LFS indicators
}
\examples{
# Using the data available to this dashboard
\dontrun{
dash_data <- load_dash_data()
table_ids <- c(
  "A84423349V",
  "A84423356T",
  "A84423355R",
  "A84423354L",
  "A84423350C",
  "A85223451R"
)

table_data <- filter_dash_data(table_ids)

table_overview()
}

}
\author{
Darren Wong, Duy Nguyen
}
