% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_regions.R
\name{title_unemprate_vic}
\alias{title_unemprate_vic}
\alias{title_unemp_emppop_partrate_vic}
\title{Function to create the graphs for the 'Regions' subpage on the dashboard.}
\usage{
title_unemp_emppop_partrate_vic(
  data = filter_dash_data(c("A84599659L", "A84600019W", "A84600187J", "A84599557X",
    "A84600115W", "A84599851L", "A84599923L", "A84600025T", "A84600193C", "A84599665J",
    "A84600031L", "A84599671C", "A84599677T", "A84599683L", "A84599929A", "A84600121T",
    "A84600037A", "A84599658K", "A84599660W", "A84600018V", "A84600020F", "A84600186F",
    "A84600188K", "A84599556W", "A84599558A", "A84600114V", "A84600116X", "A84599850K",
    "A84599852R", "A84599922K", "A84599924R", "A84600024R", "A84600026V", "A84600192A",
    "A84600194F",      "A84599664F", "A84599666K", "A84600030K", "A84600032R",
    "A84599670A", "A84599672F", "A84599676R", "A84599678V", "A84599682K", "A84599684R",
    "A84599928X", "A84599930K", "A84600120R", "A84600122V", "A84600036X", "A84600038C"),
    df = dash_data),
  selected_indicator = "unemp_rate"
)
}
\arguments{
\item{data}{the dataframe containing data to visualise}
}
\description{
Function to create the graphs for the 'Regions' subpage on the dashboard.
}
\examples{
\dontrun{

dash_data <- load_dash_data()

# for 'viz_emp_regions_sincecovid':
data <- filter_dash_data(c(
  "A84600141A",
  "A84600075R"
)) \%>\%
  dplyr::filter(date >= as.Date("2020-01-01"))

# for 'viz_reg_unemprate_multiline':
data <- filter_dash_data(c(
  "A84600253V",
  "A84599659L",
  "A84600019W",
  "A84600187J",
  "A84599557X",
  "A84600115W",
  "A84599851L",
  "A84599923L",
  "A84600025T",
  "A84600193C",
  "A84599665J",
  "A84600031L",
  "A84599671C",
  "A84599677T",
  "A84599683L",
  "A84599929A",
  "A84600121T",
  "A84600037A"
))
}
}
