% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_summarise.R
\name{ts_summarise}
\alias{ts_summarise}
\title{Summarise a data frame containing ABS time series}
\usage{
ts_summarise(df)
}
\arguments{
\item{df}{Data frame containing time series}
}
\value{
List with
\itemize{
\item This is an item.
}
}
\description{
Summarise a data frame containing ABS time series
}
