% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_update-data-and-knit.R
\name{update_data}
\alias{update_data}
\title{Update Package Data and Knit Report}
\usage{
update_data()
}
\value{

}
\description{
This Rstudio addin will run the code to update data, and
clean the app-cache
}
