% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_update-data-and-knit.R
\name{update_data_and_knit}
\alias{update_data_and_knit}
\title{Update Package Data and Knit Report}
\usage{
update_data_and_knit()
}
\description{
This Rstudio addin will run the code to update data,
clean the app-cache and then render the Monthly report
}
