% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_industries.R
\name{vic_industries_employment_treemap}
\alias{vic_industries_employment_treemap}
\title{Function to create the graphs for the 'Industries' subpage on the dashboard.}
\usage{
vic_industries_employment_treemap(
  data = filter_dash_data(c("A84601680F", "A84601683L", "A84601686V", "A84601665J",
    "A84601704L", "A84601707V", "A84601710J", "A84601638A", "A84601653X", "A84601689A",
    "A84601656F", "A84601713R", "A84601668R", "A84601695W", "A84601698C", "A84601650T",
    "A84601671C", "A84601641R", "A84601716W", "A84601662A", "A84601681J", "A84601684R",
    "A84601687W", "A84601666K", "A84601705R", "A84601708W", "A84601711K", "A84601639C",
    "A84601654A", "A84601690K", "A84601657J", "A84601714T", "A84601669T", "A84601696X",
    "A84601699F",      "A84601651V", "A84601672F", "A84601642T", "A84601717X",
    "A84601663C", "A84601682K", "A84601685T", "A84601688X", "A84601667L", "A84601706T",
    "A84601709X", "A84601712L", "A84601640L", "A84601655C", "A84601691L", "A84601658K",
    "A84601715V", "A84601670A", "A84601697A", "A84601700C", "A84601652W", "A84601673J",
    "A84601643V", "A84601718A", "A84601664F"), df = dash_data)
)
}
\arguments{
\item{data}{the dataframe containing data to visualise}
}
\description{
Function to create the graphs for the 'Industries' subpage on the dashboard.
}
\examples{
\dontrun{

dash_data <- load_dash_data()

# for 'viz_industries_empchange_sincecovid_bar':
data <- filter_dash_data(c(
  "A84601680F",
  "A84601683L",
  "A84601686V",
  "A84601665J",
  "A84601704L",
  "A84601707V",
  "A84601710J",
  "A84601638A",
  "A84601653X",
  "A84601689A",
  "A84601656F",
  "A84601713R",
  "A84601668R",
  "A84601695W",
  "A84601698C",
  "A84601650T",
  "A84601671C",
  "A84601641R",
  "A84601716W",
  "A84601662A"
)) \%>\%
  dplyr::filter(date >= as.Date("2020-01-01"))
}
}
