% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{viz_emptotal_vic}
\alias{viz_emptotal_vic}
\title{Function to make graph 'I-1-1', which is the first graph on the Indicators page
x = employed Victorians, seasonally adjusted '000
y = months, starting from}
\usage{
viz_emptotal_vic(data)
}
\description{
Function to make graph 'I-1-1', which is the first graph on the Indicators page
x = employed Victorians, seasonally adjusted '000
y = months, starting from
}
