% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_groups.R
\name{viz_gr_gen_emp_bar}
\alias{viz_gr_gen_emp_bar}
\title{Function to create the graphs for the 'Groups' subpage on the dashboard.}
\usage{
viz_gr_gen_emp_bar(
  data = filter_dash_data(c("A84423349V", "A84423237A", "A84423461V", "A84423357V",
    "A84423245A", "A84423469L", "A84423350C", "A84423238C", "A84423462W"), df =
    dash_data),
  title = ""
)
}
\arguments{
\item{data}{the dataframe containing data to visualise}
}
\description{
Function to create the graphs for the 'Groups' subpage on the dashboard.
}
