% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_industries.R
\name{viz_industries_empchange_sincecovid_bar}
\alias{viz_industries_empchange_sincecovid_bar}
\title{Function to create the graphs for the 'Industries' subpage on the dashboard.}
\usage{
viz_industries_empchange_sincecovid_bar(
  data = filter_dash_data(c("A84601680F", "A84601683L", "A84601686V", "A84601665J",
    "A84601704L", "A84601707V", "A84601710J", "A84601638A", "A84601653X", "A84601689A",
    "A84601656F", "A84601713R", "A84601668R", "A84601695W", "A84601698C", "A84601650T",
    "A84601671C", "A84601641R", "A84601716W", "A84601662A"), df = dash_data)
)
}
\arguments{
\item{data}{the dataframe containing data to visualise}
}
\description{
Function to create the graphs for the 'Industries' subpage on the dashboard.
}
\examples{
\dontrun{

dash_data <- load_dash_data()

# for 'viz_industries_empchange_sincecovid_bar':
data <- filter_dash_data(c(
  "A84601680F",
  "A84601683L",
  "A84601686V",
  "A84601665J",
  "A84601704L",
  "A84601707V",
  "A84601710J",
  "A84601638A",
  "A84601653X",
  "A84601689A",
  "A84601656F",
  "A84601713R",
  "A84601668R",
  "A84601695W",
  "A84601698C",
  "A84601650T",
  "A84601671C",
  "A84601641R",
  "A84601716W",
  "A84601662A"
)) \%>\%
  dplyr::filter(date >= as.Date("2020-01-01"))
}
}
