% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_words.R
\name{combine_words}
\alias{combine_words}
\title{Combine elements of vector into a single string with commas and 'and'}
\usage{
combine_words(words)
}
\arguments{
\item{words}{Vector of element(s) to be combined into one string.}
}
\value{
A length one character vector; elements will be separated by commas.
}
\description{
Simplified version of \code{knitr::combine_words()}
}
\examples{
combine_words(c("a", "b", "c"))
combine_words(c(1, 2, 3))
}
\author{
knitr authors
}
