% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_df.R
\name{create_summary_df}
\alias{create_summary_df}
\title{Create a summary dataframe for use in a table to be displayed
Either in a reactable or a static briefing table}
\usage{
create_summary_df(
  data,
  dashboard_or_briefing = "dashboard",
  years_in_sparklines = 2
)
}
\arguments{
\item{data}{Dataframe of input data}

\item{dashboard_or_briefing}{Character; either 'dashboard'
(if the table is intended for the dashboard),
or 'briefing' (if the table is intended for the briefing tables).}

\item{years_in_sparklines}{Number of years worth of data to include in
sparkline}
}
\description{
Create a summary dataframe for use in a table to be displayed
Either in a reactable or a static briefing table
}
\examples{
\dontrun{
dash_data <- load_dash_data()
create_summary_df(data = filter_dash_data(c(
  "A84423349V",
  "A84423357V",
  "A84423356T",
  "A84423244X",
  "A84423468K",
  "pt_emp_vic"
)))
}
}
