% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_df.R
\name{create_summary_df}
\alias{create_summary_df}
\title{Create a summary dataframe for use in a table to be displayed
Either in a reactable or a static briefing table}
\usage{
create_summary_df(
  data,
  for_reactable = TRUE,
  pretty_names = FALSE,
  years_in_sparklines = 2,
  row_var = "indicator"
)
}
\arguments{
\item{data}{Dataframe of input data}

\item{for_reactable}{logical; if TRUE, series_id column is retained,
data for sparkline is included, column names are not nicely formatted}

\item{row_var}{"Quoted" name of column that contains series names to be
used as the row names in the output table}

\item{years_in_sparkline}{Number of years worth of data to include in
sparkline}
}
\description{
Create a summary dataframe for use in a table to be displayed
Either in a reactable or a static briefing table
}
