% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_df.R
\name{create_summary_df}
\alias{create_summary_df}
\title{Create a summary dataframe for use in a table to be displayed
Either in a reactable or a static briefing table}
\usage{
create_summary_df(
  data,
  for_reactable = TRUE,
  pretty_names = FALSE,
  years_in_sparklines = 2,
  row_var = "indicator",
  period = "month",
  row_order = NULL
)
}
\arguments{
\item{data}{Dataframe of input data}

\item{for_reactable}{logical; if TRUE, series_id column is retained,
data for sparkline is included, column names are not nicely formatted}

\item{row_var}{"Quoted" name of column that contains series names to be
used as the row names in the output table}

\item{period}{Period covered by most recent change, such as "month" or
"quarter". Used for column heading in the returned DF.}

\item{row_order}{If \code{NULL}, table will be sorted in alphabetical order.
Otherwise, \code{row_order} should be a vector of row names in the order in which
you want them to appear in the table.}

\item{years_in_sparkline}{Number of years worth of data to include in
sparkline}
}
\description{
Create a summary dataframe for use in a table to be displayed
Either in a reactable or a static briefing table
}
\examples{
create_summary_df(data = filter_dash_data(c(
  "A84423349V",
  "A84423357V",
  "A84423356T",
  "A84423244X",
  "A84423468K",
  "pt_emp_vic"
)))
}
