% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_briefing_table.R
\name{make_briefing_table}
\alias{make_briefing_table}
\title{Take a data frame summarising data and create a flextable for
a briefing table}
\usage{
make_briefing_table(df, highlight_rows = NULL)
}
\arguments{
\item{df}{Dataframe created using \code{create_summary_df()}}

\item{highlight_rows}{numeric vector of rows in the table body (ie.
excluding column names / header row) to highlight. Non-highlighted rows
will be indented. When \code{NULL}, no row is highlighted.}
}
\value{
A \code{flextable::flextable} object
}
\description{
Take a data frame summarising data and create a flextable for
a briefing table
}
\examples{
dash_data <- load_dash_data()

overview_data <- filter_dash_data(series_ids = c(
  "A84423349V", # Employed total
  "A84423355R", # Part rate
  "A84423354L", # Unemp rate
  "A84423350C", # Unemp total
  "A85223451R", # Underut rate
  "A84426256L", # Hours worked
  "A85223450L", # Underemp rate
  "A84423357V", # Emp FT
  "pt_emp_vic" # Emp PT
))

x <- create_summary_df(overview_data, for_reactable = FALSE)

make_briefing_table(x)
}
