% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_reactable.R
\name{make_reactable}
\alias{make_reactable}
\title{Make a reactable with standard formatting}
\usage{
make_reactable(
  data,
  years_in_sparklines = 2,
  row_var = "indicator",
  row_order = NULL
)
}
\arguments{
\item{data}{a data frame}

\item{years_in_sparklines}{number of years prior to latest obs to include
in sparkline}

\item{row_var}{"Quoted" name of column in \code{data} to use as the row names in
the table - as in "indicator"}

\item{row_order}{If \code{NULL}, table will be sorted in alphabetical order.
Otherwise, \code{row_order} should be a vector of row names in the order in which
you want them to appear in the table.}
}
\description{
Each row is an indicator; columns are levels / change over particular periods
Conditional formatting is applied, sparklines are included.
}
