% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\title{Create a table for the dashboard or a briefing document}
\usage{
make_table(
  data,
  dashboard_or_briefing = "dashboard",
  years_in_sparklines = 2,
  row_order = NULL,
  highlight_rows = NULL
)
}
\arguments{
\item{data}{A data frame containing data to summarise}

\item{dashboard_or_briefing}{either "dashboard" or "briefing", specifying
the destination for the output table}

\item{years_in_sparklines}{Period of time to include in the sparkline line
charts. Only relevant when \code{dashboard_or_briefing} =
"dashboard".}

\item{row_order}{Vector of series IDs, in the order in which you wish the
corresponding rows to be included in the output table}

\item{highlight_rows}{Numeric vector of rows in the table to highlight.
Highlighted rows are bolded and have a top border; non-highlighted rows
are indented. If \code{NULL} then all rows are non-bold, non-indented.}
}
\value{
Either a \code{reactable} (if \code{dashboard_or_briefing} = "dashboard") or a
\code{flextable} (if \code{dashboard_or_briefing} = "briefing") summarising \code{data}.
}
\description{
Create a table for the dashboard or a briefing document
}
\examples{
\dontrun{
dash_data <- load_dash_data()

make_table(
data = filter_dash_data(series_ids = c(
  "A84423354L",
    "A84423242V",
 "A84423466F",
 "A84433601W",
 "A84600079X",
 "A84423350C",
 "A84423349V",
 "A84423357V",
 "pt_emp_vic",
 "A84423461V",
 "A84423237A",
 "A84424687C",
 "A84423355R",
 "A84423243W",
 "A84423467J",
 "A84433602X",
 "A84426256L",
 "A85223450L",
 "A85223451R",
 "A84423356T"
)),
dashboard_or_briefing = "dashboard",
row_order = c(
 "A84423354L",
 "A84423242V",
 "A84423466F",
 "A84433601W",
 "A84600079X",
 "A84423350C",
 "A84423349V",
 "A84423357V",
 "pt_emp_vic",
 "A84423461V",
 "A84423237A",
 "A84424687C",
 "A84423355R",
 "A84423243W",
 "A84423467J",
 "A84433602X",
 "A84426256L",
 "A85223450L",
 "A85223451R",
 "A84423356T"
),
highlight_rows = c(1, 6, 7, 13, 17, 18, 19, 20)
)

}
}
