% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_date.R
\name{reference_dates}
\alias{reference_dates}
\title{Labour Force Reference Dates}
\usage{
reference_dates(detailed = FALSE)
}
\arguments{
\item{detailed}{logical standard or detailed Labour Force release}
}
\value{
list
}
\description{
Get Labour Force reference dates and generate email preamble
}
