% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_indicators.R
\name{viz_empgrowth_sincecovid}
\alias{viz_empgrowth_sincecovid}
\title{Line chart of cumulative employment change since March 2020
in Victoria and Australia}
\usage{
viz_empgrowth_sincecovid(data)
}
\description{
Line chart of cumulative employment change since March 2020
in Victoria and Australia
}
\examples{
\dontrun{
dash_data <- load_dash_data()
df <- filter_dash_data(c("A84423043C", "A84423349V"), dash_data)
viz_empgrowth_sincecovid(df)
}

}
