% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_groups.R
\name{viz_gr_gen_emp_bar}
\alias{viz_gr_gen_emp_bar}
\title{Function to create the graphs for the 'Groups' subpage on the dashboard.}
\usage{
viz_gr_gen_emp_bar(
  data = filter_dash_data(c("A84423469L", "A84423245A", "A84423801C", "A84423577W",
    "A84423461V", "A84423237A", "A84423463X", "A84423239F", "A84423462W", "A84423238C"),
    df = dash_data) \%>\% dplyr::group_by(.data$series) \%>\% dplyr::filter(.data$date ==
    max(.data$date)),
  title = ""
)
}
\arguments{
\item{data}{the dataframe containing data to visualise}
}
\description{
Function to create the graphs for the 'Groups' subpage on the dashboard.
}
\examples{
\dontrun{

# for viz_gr_gen_emp_bar:
ids <- c(
  "A84423349V",
  "A84423237A",
  "A84423461V",
  "A84423357V",
  "A84423245A",
  "A84423469L",
  "A84423350C",
  "A84423238C",
  "A84423462W"
)
# (also parttime figures and
# not in the labour force)

# for viz_gr_gen_partrate_line:
ids <- c(
  "A84423355R",
  "A84423243W",
  "A84423467J"
)

# for viz_gr_gen_unemp_line:
ids <- c(
  "A84423354L",
  "A84423242V",
  "A84423466F"
)

# for viz_gr_yth_lfpart_line:
# ids <- will need smoothing and extracting data from LM1.

# for viz_gr_yth_emp_line:
}
}
